requirements.txt
asgiref==3.2.10
cffi==1.14.0
cloudpickle
cycler==0.10.0
dask
decorator
Django==3.0.8
future==0.18.2
image==1.5.32
imageio
joblib==0.16.0
kiwisolver
matplotlib
networkx
numpy==1.19.0
pandas
Pillow==7.2.0
pycparser
pygit==0.1
pyparsing
python-dateutil
pytz
PyWavelets==1.1.1
PyYAML
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
six
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.7.4
toolz
tqdm
imageio-ffmpeg
moviepy